+++ common/components/Forms/AddPaymentForm/PaymentPricesTable/fields/priceFields.tsx
@@ -5,9 +5,6 @@ import s from '../style.module.scss'
 import useCompany from '@common/modules/hooks/useCompany'
 import useService from '@common/modules/hooks/useService'
 import { usePaymentContext } from '@common/components/AddPaymentModal'
-import { useCompanyInvoice } from '@common/modules/hooks/usePayment'
-import { NamePath } from 'antd/lib/form/interface'
-import { ServiceType } from '@utils/constants'
 import { useInflicionValues } from './amountFields'
 
 export function PriceMaintainceField({ record, edit }) {
@@ -80,66 +77,6 @@ function InflicionPricePlacingField({ record, edit }) {
   )
 }
 
-export function OldElectricity({ record, edit }) {
-  return (
-    <FormAttributeForSingle
-      lastAmountName={[record.name, 'lastAmount']}
-      invoicePropName={ServiceType.Electricity}
-      amountName={[record.name, 'amount']}
-      disabled={edit}
-    />
-  )
-}
-
-export function OldWater({ record, edit }) {
-  return (
-    <FormAttributeForSingle
-      lastAmountName={[record.name, 'lastAmount']}
-      amountName={[record.name, 'amount']}
-      invoicePropName={ServiceType.Water}
-      disabled={edit}
-    />
-  )
-}
-
-function FormAttributeForSingle({
-  invoicePropName,
-  lastAmountName,
-  amountName,
-  disabled,
-}: {
-  invoicePropName: string
-  lastAmountName: NamePath
-  amountName: NamePath
-  disabled?: boolean
-}) {
-  const { paymentData, form } = usePaymentContext()
-  const companyId = Form.useWatch('company', form) || paymentData?.company
-  const { lastInvoice } = useCompanyInvoice({ companyId, skip: disabled })
-  const value = lastInvoice?.invoice?.find(
-    (item) => item.type === invoicePropName
-  )?.amount
-
-  useEffect(() => {
-    if (!disabled) {
-      form.setFieldValue(lastAmountName, value)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [value])
-
-  return (
-    <div className={s.doubleInputs}>
-      <Form.Item name={lastAmountName} rules={validateField('required')}>
-        <InputNumber disabled={disabled} className={s.input} />
-      </Form.Item>
-
-      <Form.Item name={amountName} rules={validateField('required')}>
-        <InputNumber disabled={disabled} className={s.input} />
-      </Form.Item>
-    </div>
-  )
-}
-
 export function PriceElectricityField({ record, edit }) {
   const { paymentData, form } = usePaymentContext()
   const fieldName = [record.name, 'price']
@@ -200,10 +137,12 @@ export function PriceGarbageCollectorField({ record, edit }) {
       company?.garbageCollector &&
       service?.garbageCollectorPrice
     ) {
-      form.setFieldValue(fieldName, service.garbageCollectorPrice)
+      form.setFieldValue(
+        fieldName,
+        (service.garbageCollectorPrice / 100) * company.rentPart
+      )
     }
-  }, [service?._id, service?.garbageCollectorPrice]) //eslint-disable-line react-hooks/exhaustive-deps
-
+  }, [service?._id, company?.garbageCollector]) //eslint-disable-line react-hooks/exhaustive-deps
   return (
     <Form.Item name={fieldName} rules={validateField('required')}>
       <InputNumber disabled={edit} className={s.input} />
@@ -266,43 +205,6 @@ export function PriceWaterPartField({ record, edit }) {
     <Form.Item name={fieldName} rules={validateField('required')}>
       <InputNumber disabled={edit} className={s.input} />
     </Form.Item>
-  )
-}
-
-export function WaterPartInfo({ edit }) {
-  const { paymentData, form } = usePaymentContext()
-  const serviceId =
-    Form.useWatch('monthService', form) || paymentData?.monthService
-  const companyId = Form.useWatch('company', form) || paymentData?.company
-  const { company } = useCompany({ companyId, skip: edit })
-  const { service } = useService({ serviceId, skip: edit })
-
-  return (
-    <>
-      {company?.waterPart && service?.waterPriceTotal
-        ? company.waterPart + '% від суми ' + service?.waterPriceTotal
-        : null}
-    </>
-  )
-}
-
-export function InflicionAmountInfo({ edit }) {
-  const { paymentData, form } = usePaymentContext()
-  const serviceId =
-    Form.useWatch('monthService', form) || paymentData?.monthService
-  const companyId = Form.useWatch('company', form) || paymentData?.company
-  const { company } = useCompany({ companyId, skip: edit })
-  const { service } = useService({ serviceId, skip: edit })
-
-  const percent = (service?.inflicionPrice - 100).toFixed(2)
-  const rentPrice = company?.pricePerMeter * company?.totalArea
-
-  return (
-    <>
-      {company?.inflicion && service?.inflicionPrice
-        ? percent + '% інфляції від ' + rentPrice
-        : null}
-    </>
   )
 }
