+++ utils/testData.ts
@@ -104,7 +104,7 @@ export const realEstates = [
     description: 'none',
     servicePricePerMeter: 10,
     totalArea: 10,
-    garbageCollector: 11,
+    garbageCollector: true,
     waterPart: 0,
     rentPart: 0,
     inflicion: false,
@@ -119,7 +119,7 @@ export const realEstates = [
     servicePricePerMeter: 10,
     description: 'none',
     totalArea: 10,
-    garbageCollector: 12,
+    garbageCollector: true,
     waterPart: 0,
     rentPart: 0,
     inflicion: false,
@@ -134,7 +134,7 @@ export const realEstates = [
     pricePerMeter: 10,
     servicePricePerMeter: 10,
     totalArea: 10,
-    garbageCollector: 13,
+    garbageCollector: true,
     waterPart: 0,
     rentPart: 0,
     inflicion: false,
@@ -149,7 +149,7 @@ export const realEstates = [
     servicePricePerMeter: 10,
     description: 'none',
     totalArea: 10,
-    garbageCollector: 14,
+    garbageCollector: true,
     waterPart: 0,
     rentPart: 0,
     inflicion: false,
@@ -164,7 +164,7 @@ export const realEstates = [
     pricePerMeter: 10,
     servicePricePerMeter: 10,
     totalArea: 10,
-    garbageCollector: 13,
+    garbageCollector: true,
     waterPart: 0,
     rentPart: 0,
     inflicion: false,
@@ -179,7 +179,7 @@ export const realEstates = [
     pricePerMeter: 10,
     servicePricePerMeter: 10,
     totalArea: 10,
-    garbageCollector: 13,
+    garbageCollector: true,
     waterPart: 0,
     rentPart: 0,
     inflicion: false,
